refactor(operator): Split Kubernetes operator module
Refactored operator module to follow single responsibility principle:

**Module Structure (312 lines → 6 files):**
- types.rs (41 lines): OperatorError, Context, Result, FINALIZER_NAME
- resources.rs (132 lines): Deployment/Service creation functions
- mcpserver.rs (170 lines): MCPServer controller and reconciliation logic
- plugin.rs (43 lines): Plugin controller
- security.rs (46 lines): SecurityPolicy controller
- mod.rs (15 lines): Re-exports only

**Benefits:**
- Clear separation of concerns (types, resources, controllers)
- Each controller in its own file for better maintainability
- Resource creation logic isolated in dedicated module
- Easier to extend with new controllers

**Testing:**
- All 377 tests passing
- No clippy warnings
- Formatting check passed
- Public API unchanged

Closes #163
